Migrate docs gulp scripts to node scripts
