+++ packages/node-build-scripts/generate-docs-data
@@ -1,0 +1,100 @@
+#!/usr/bin/env node
+/**
+ * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
+ * @fileoverview Generates data for packages/docs-app
+ */
+
+const { spawn } = require("child_process");
+const dm = require("documentalist");
+const fs = require("fs");
+const glob = require("glob");
+const path = require("path");
+const semver = require("semver");
+const docsUtils = require("./docsUtils");
+
+// assume we are running from packages/docs-app
+const PACKAGES_DIR = path.resolve(process.cwd(), "../");
+const GENERATED_SRC_DIR = path.resolve(process.cwd(), "./src/generated");
+const BUILD_DIR = path.resolve(process.cwd(), "build");
+const NAV_PAGE_NAME = "_nav";
+
+const DOCS_DATA_FILENAME = "docs.json";
+const DOCS_RELEASES_FILENAME = "releases.json";
+const DOCS_VERSIONS_FIELENAME = "versions.json";
+
+generateDocumentalistData();
+generateReleasesData();
+generateVersionsData();
+
+/**
+ * Run documentalist to generate docs data from source code.
+ */
+function generateDocumentalistData() {
+    new dm.Documentalist({
+        markdown: { renderer: docsUtils.markedRenderer },
+        // must mark our @Decorator APIs as reserved so we can use them in code samples
+        reservedTags: ["import", "ContextMenuTarget", "HotkeysTarget"],
+    })
+        .use(".md", new dm.MarkdownPlugin({ navPage: NAV_PAGE_NAME }))
+        .use(/\.d\.ts$/, new dm.TypescriptPlugin({
+            excludeNames: [/Factory$/, /^I.+State$/],
+            excludePaths: ["node_modules/"],
+            includeDefinitionFiles: true,
+        }))
+        .use(".scss", new dm.KssPlugin())
+        .documentGlobs("../*/src/**/*", "../*/dist/index.d.ts")
+        .then((docs) => JSON.stringify(docs, null, 2))
+        .then((content) => (
+            fs.writeFileSync(path.join(GENERATED_SRC_DIR, DOCS_DATA_FILENAME), content)
+        ));
+}
+
+/**
+ * Create a JSON file containing latest released version of each project
+ */
+function generateReleasesData() {
+    const packageDirectories = fs.readdirSync(PACKAGES_DIR)
+        .map(name => path.join(PACKAGES_DIR, name))
+        .filter(source => fs.lstatSync(source).isDirectory());
+
+    const releases = packageDirectories
+        .map((packagePath) => require(path.resolve(packagePath, "package.json")))
+        // only include non-private projects
+        .filter((project) => !project.private)
+        // just these two fields from package.json:
+        .map(({ name, version }) => ({ name, version }));
+
+    fs.writeFileSync(path.join(GENERATED_SRC_DIR, DOCS_RELEASES_FILENAME), JSON.stringify(releases, null, 2));
+}
+
+/**
+ * Create a JSON file containing published versions of the documentation
+ */
+function generateVersionsData() {
+    let stdout = "";
+    const child = spawn("git", ["tag"]);
+    child.stdout.setEncoding("utf8");
+    child.stdout.on("data", data => { stdout += data; });
+    child.on("close", () => {
+        /** @type {Map<string, string>} */
+        const majorVersionMap = stdout.split("\n")
+            // turn release-* tags into version numbers
+            .filter(val => /release-[1-9]\d*\.\d+\.\d+.*/.test(val))
+            .map(val => val.slice(8))
+            // inject current version (unreleased package bump)
+            .concat(require(path.resolve(process.cwd(), "package.json")).version)
+            .reduce((map, version) => {
+                const major = semver.major(version);
+                if (!map.has(major) || semver.gt(version, map.get(major))) {
+                    map.set(major, version);
+                }
+                return map;
+            }, new Map());
+        // sort in reverse order (so latest is first)
+        const majorVersions = Array.from(majorVersionMap.values()).sort(semver.rcompare);
+
+        console.log("[generate-docs-data] Versions:", majorVersions.join(", "));
+
+        fs.writeFileSync(path.join(GENERATED_SRC_DIR, DOCS_VERSIONS_FIELENAME), JSON.stringify(majorVersions, null, 2));
+    });
+}
